feat: subaccounts (#431)
* Implement GetSubAccountsRequest

* Fix use case for GetSubAccounts

* Add XmlDoc for SubAccount

* Rename SubAccount record to Account, as it can also represent the primary account

* Implement GetSubAccountRequest, with internal WithApiKey method

* Implement GetSubAccount

* Implement CreateSubAccountRequest

* Revert update on solution with temp project

* Fix smells

* Implement UseCase for CreateSubAccount

* Implement UpdateSubAccountRequest

* Fix console app project

* Remove console app completely

* Implement TransferCredit

* Implement GetCreditTransfers

* Group CreditTransfer and BalanceTransfer into same record

* Merge use cases for CreditTransfer and BalanceTransfer

* Implement TransferNumber

* Fix url path in requests

* Unify Transfer for credit/balance

* Group GetCreditTransfers and GetBalanceTransfers together

* Reorganize tests for GetTransfers use cases

* Remove duplicate method on client

* Flag current implementation as obsolete

+++ Vonage.Common/Validation/InputValidation.cs
@@ -168,7 +168,7 @@ public static class InputValidation
     /// <param name="name">The display name.</param>
     /// <typeparam name="T">The request type.</typeparam>
     /// <returns>Success or Failure.</returns>
-    public static Result<T> VerifyNotNegative<T>(T request, int value, string name) =>
+    public static Result<T> VerifyNotNegative<T>(T request, decimal value, string name) =>
         value < 0
             ? Result<T>.FromFailure(
                 ResultFailure.FromErrorMessage($"{name} {IntCannotBeNegative}"))
